first publishing of ffDiaporama 1.4 devel (32bits improvements & mmfiler inclusion)
git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@311 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/src/ffDiaporama/DlgImage/cImgInteractiveZone.cpp
@@ -182,7 +182,7 @@ void cImgInteractiveZone::RefreshDisplay() {
     if (width()<height()) dmax=width(); else dmax=height();
 
     QImage  *SourceImage=NULL;
-    double  Hyp         =sqrt(CachedImage->width()*CachedImage->width()+CachedImage->height()*CachedImage->height());   // Calc hypothenuse of the image to define full canvas
+    double  Hyp         =sqrt(double(CachedImage->width())*double(CachedImage->width())+double(CachedImage->height())*double(CachedImage->height()));   // Calc hypothenuse of the image to define full canvas
 
     // calc rectangle before rotation
     double  rx=double(CachedImage->width())*(dmax/Hyp)/2;
